requirements.txt
streamlit
openai
chromadb
langchain
python-dotenv
pypdf
